Add the original SVG
It was downloaded from https://gist.github.com/johan/1007813; according
to the owner of the Gist, Johan Sundström, it was converted from
GitHub's original media in Illustrator format. While the provided link
does not point to it any longer, it is probably the same as here:
https://github.com/github/media/tree/master/octocats

There is a link in that latter repository to the page
https://github.com/logos which unfortunately contradicts the text in
http://octodex.github.com/faq.html: while the FAQ in the latter states:

	Q: Can I use an octocat as my avatar?

	You can use an octocat as your personal avatar, but not for your
	company or a product you're building. You are welcome to show
	your <3 for GitHub, but not appear as if you == GitHub.

the text in the former says:

	Please don’t do these things
	...
	Create a modified version of the Octocat or GitHub logo

Therefore, let's be clear about the purpose of my octocats: they are
*just* a nice pastime for myself. You can look at them, but that's
pretty much about it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>